+++ packages/reef/src/reef/plugin.py
@@ -39,6 +39,7 @@ def reef(ctx: Context, opts: ReefPluginOptions):
 
 	logger.debug("Created reef:item/element_base")
 
+	ctx.require(assets.element)
 	ctx.require(assets.pdf)
 	ctx.require(data.special)
 	ctx.require(data.page)